Read i/p data from local file instead of web site.
Changed code & local file structure to accomplish this change.

Use a for loop to debug the extract_one_letter_to_editor() code,
by limiting it to one or two passes of the data.  This will help
until I fix the code that discovers the number of headings and
letters and can loop on the real data.