Clear the author details array after each letter.
This code fixes the problem where details about a prior author
would appear in that array for a subsequent author (when they have
letters under the same topic).

It should also be useful to indicate when a new letter
has been started, which has been difficult to discern before
now.

The fix: compare the current & prior author's names
each time we enter the branch that handles author data.