Add leading zeroes to strings for dir & file names.
This change makes the names based on month and day of month
and a letter's sequence number, into 2-character strings instead
of either one or two.  This makes for consistent appearance and
correct sequence in a sorted list.

Add $author_count & use it with $total_letters_count to indicate
how many items were seen & handled.  This code needs improvement,
and using this variable name will clarify
what is being printed.  The two separate vars also help with adding
leading zeroes appropriately, while keeping the total counter
correct.

Add $current_line and $prior_line, and plan to use them to help
to determine the end of a letter, to fix a bug (see b.10 and
b.12).  This code is not working yet.