Interim commit during debug; not yet working.
Check the input for lines matching anchor tags, the <a pattern.
Found that this is not adequate to identify the end of each letter
in a topic.

Also found that clearing $current_author at various places
in the code creates other problems while not fixing any problem
completely, so that was abandoned.

Next idea: read a window of 3 or more lines at a time, and use
them to identify end of a letter.