Two loop code fixes bug of mangled letters.
Two letters were mangled when the second letter
had italic tags <i> in the letter body.
That is not seen in the o/p files now.

Other bugs are now seen: author details are not handled properly,
and some of them appear in the following o/p file.
Also, the topic is in the author details field & it should not
be there.