Use author name as hash key to store multiple authors per letter.
Changed from key = 'author'
          to key = the author's real name.

Thus, each author of a letter has a hash, keyed on their name.

This change to the data structure works, for the test case I have
w/ 2 authors for 1 letter.  Store each author's data in the hash
w/ that author's name as key to the hash.  The values include
author's name, and an array of data about the author, eg, title,
organization, location.

Will this be a problem if two authors have the same name?