Add fields to json o/p file: web_page_date, source_id, category.
Category LTTE: letters to the editor.

Use serverTime from web page for the web_page_date, which may
not always be the same as the date it appears in print.

source_id is a short, unique abbreviation for the source; the
only one now is wsj for The Wall Street Journal.