Use the program's --test option to run this test.
Test mode forces the program to use a local file for i/p data,
and not query the web site.  This ensures consistent operation
for this test: the same i/p should produce o/p that matches the
reference data stored on disk.  If not, the program has a problem.