Add -q cmd line option; make default = print o/p to screen.
Make it easy for casual users to get the web page's data on
their screen when running the program, by making that the default
behavior.  Replace the use of the --verbose option (and the
default of not showing output) w/ the --quiet option, that
prevents the letters from being shown on the screen.